requirements.txt
click==6.7
Flask==1.0.2
Flask-Uploads==0.2.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
Werkzeug==0.14.1
